# Skip intermediary Any when needing an interface anyway